refactor: rename previous commits to relevant commits
